# Replace forwardable with explicit delegation for performance