@koic 2024-08-30
This might be sufficient if the abstraction provided for unknown methods is not needed.