@marcandre 2024-08-30
So indeed it looks like `forwardable` [executes extra code](https://github.com/ruby/ruby/blob/master/lib/forwardable.rb#L203) than what is necessary.

1) Has there ever been a feature request opened on ruby-core for a `minimal: true` option or similar to `def_delegate`?
2) Why don't we instead make our own `forwardable` clone that only does the forwarding?