@Earlopain 2024-08-30
> This might be sufficient if the abstraction provided for unknown methods is not needed.

Can you elaborate? I don't know what you mean with that.

> Has there ever been a feature request...

I searched through the bugtracker and haven't found something. It did remind me that ruby has thoughs about making `forwardable` a bundled gem though... I commented on the issue for that. Not a fan

> Why don't we instead make our own...

I honestly don't see much value in that. Few places really _need_ it, delegating numerous methods. Most just do one or two which I think is fine to expliciy type out. But if you insist I will amend the PR.