@marcandre 2024-08-30
> suffici



> > Why don't we instead make our own...
> 
> I honestly don't see much value in that. Few places really _need_ it, delegating numerous methods. Most just do one or two which I think is fine to expliciy type out. But if you insist I will amend the PR.

I do find the older code clearer than the new one. It was using a simple abstraction explicitly. We already have the code (the `class_eval`), we just need to extract it to a helper function.