@Earlopain 2024-08-30
> I do find the older code clearer

That's fair, I don't feel too strongly myself. I will amend the ast PR later when I got a bit more time and offer something similar here