@Earlopain 2024-08-30
Please do, you probably have something in mind already. I'm not sure where to put. You should be able to edit my branch if you want, or just make it from scratch if you prefer.